Footer: Every page will have a footer which is static in the users device display view port. will contain a "Control Icon Buttons" bar consisting of icon buttons for various page options/ pop ups etc.
Below this will be a "Text Input Box" which a user can input text to send to the current pages Models section, the enter/return keyboard key also auto sends to the apps selection. on the same row as the "Text Input Box" will be the "Mic" button on the left, and the "Send" button on the Right.

A "Send" Icon button to send the text that's inputted in the "Text Input Box" to the selected pages app selection. the send Icon button will be on the same row as the "Text Input Box" on the right hand side

A "Mic" Icon button to allow the user to activate the devices microphone and have the microphone dictate the users speech into text and it populates the "Text Input Box". When the "Mic" Icon button is pressed the app will ask user for device Microphone permissions for the MIc and when accepted the "Mic" Icon button will pulse red to show its active. When the user presses the active "Mic" icon button again, the Microphone is deactivated, (devices mic is turned off, cleaned and refreshed from listening), Microphones state is refreshed to off and the "Mic" Icon button stops pulsing red and reverts to its original state ready to be pressed again to reactivate if needed.

bottom row will contain a very small text banner greyed out: "Created by Jamie Reddin using Puter.com and other providers  | JR Comhrá AI © 2025 | V. 0.0.1". have the text "Jamie Reddin" as a clickable link to "https://jayreddin.github.io" and "Puter.com" as a clickable link to "https://puter.com" and "providers" as a clickable element that pops up a small card listing other clickable links: "Jamie Reddin" as a as a clickable link to "https://jayreddin.github.io/", "Jamie Discord" as a clickable link to "https://discord.gg/3YdvQfpPPr", "Puter.com" as a clickable link to "https://puter.com/", "Puter.com Discord" as a clickable link to "https://discord.gg/gtVFcCQa", "Google Gemini" as a clickable link to "https://gemini.google.com/", "OpenRouter" as a clickable link to "https://openrouter.ai/". CLicking any link opens a new tab to that link and clicking outside of the pop up container dismisses it.